Improved + fixed problems with category editing
- UI improved
- Proper use of contracts, eliminating hard coded models/requests
- Using updated category module with sorting enabled
- Added missing breadcrumbs
- UI copy standardized

+++ src/resources/views/taxon/_tree.blade.php
@@ -1,21 +1,44 @@
 @foreach($taxons as $taxon)
-    <li class='list-group-item'>
-        {{ Form::open(['url' => route('vanilo.taxon.destroy', [$taxonomy, $taxon]), 'class' => 'form', 'method' => 'DELETE']) }}
-        <div class="form-group">
+    <div class="card-block">
+        @can('edit taxons')
+            <a href="{{ route('vanilo.taxon.edit', [$taxonomy, $taxon]) }}">{{ $taxon->name }}</a>
+        @else
             {{ $taxon->name }}
+        @endcan
 
-            <span class="float-right">
-                <a class="btn btn-outline-primary btn-xs" href="{{ route('vanilo.taxon.edit', [$taxonomy, $taxon]) }}">
-                        {{ __('Edit')}}
-                </a>
-                <button type="submit" class="btn btn-outline-danger btn-xs">{{ __('Delete') }}</button>
-            </span>
-        </div>
-        {{ Form::close() }}
         @if ($taxon->children->isNotEmpty())
-            <ul class="list-group">
-                @include('vanilo::taxon._tree', ['taxons' => $taxon->children])
-            </ul>
+                <a href="#taxon-{{$taxon->id}}" aria-expanded="false"
+                   aria-controls="taxon-{{$taxon->id}}" data-toggle="collapse"
+                   class="collapse-toggler-heading">
+                    &nbsp;<i class="zmdi zmdi-chevron-right"></i>
+                </a>
         @endif
-    </li>
+        <div class="card-actionbar">
+            @can('delete taxons')
+                {{ Form::open(['url' => route('vanilo.taxon.destroy', [$taxonomy, $taxon]), 'class' => 'form', 'method' => 'DELETE']) }}
+                <button type="submit"
+                        class="btn btn-outline-danger btn-xs">{{ __('Delete') }}</button>
+                {{ Form::close() }}
+            @endcan
+        </div>
+    </div>
+
+    @if ($taxon->children->isNotEmpty())
+        <div class="collapse multi-collapse" id="taxon-{{$taxon->id}}" data-toggle="collapse">
+            <div class="card-block">
+                <div class="card">
+                    @include('vanilo::taxon._tree', ['taxons' => $taxon->children])
+                </div>
+            </div>
+        </div>
+    @endif
+
 @endforeach
+
+@can('create taxons')
+    <div class="card-footer">
+        <?php $queryParam = $taxon->isRootLevel() ? '' : '?parent=' . $taxon->parent->id; ?>
+        <a href="{{ route('vanilo.taxon.create', $taxonomy) }}{{ $queryParam }}"
+           class="btn btn-outline-success btn-sm">{{ __('Add :category', ['category' => str_singular($taxonomy->name)]) }}</a>
+    </div>
+@endcan
